Cumulative changes from main branch (trunk). With these changes ATF version is upgraded to v3.10.9.0
New Features
•	SnapToGrid (allow user to turn on/off snap to grid behavior in CircuitEditor).
•	FlagsUITypeEditor: Merged GitHub pull 58, added support for uint flags
•	BirdEyeView panel for circuit editor
•	Added new shortcut to FrameAll command.

Bug fixes
•	Circuit pin bounds was inconsistent between render code and pick code.
•	Dom: When auto converting simple element to attribute and PreserveSimpleElements is set then always write element to respect minOccurs
•	NumericTupleControl.cs: Merged Github pull 59, Fix some layout issue when resizing.
•	ViewingContext: Bug fix in GetBounds() didn't work correctly for circuit group.
•	Fixed color value in skin template that caused app to crash when creating new skin.
•	Fixed a long standing bug in fit (frame) operation.
•	D2dGraphAdapter: Bug Fix: GetBounds() transformed empty bound to client space.
•	D2dUtil: Bug in TransformVector(): it just return input without transforming it.

Breaking and Behavioral Changes:
•	Moved OSC (open sound control) to its own project
•	Direct2d: Added more error handling for recreating render target.
•	BoundedFloatEditor.cs: changed constructor's min and max types from int to floats. (They should have been float).
•	CircuitValidator.cs: Re-enabled HistoryContext's merge feature.
•	Removed Frame and EnsureVisible extension method from ITransformAdapter.

+++ Framework/Atf.Gui.WinForms/Controls/PropertyEditing/BoundedFloatEditor.cs
@@ -29,7 +29,7 @@ namespace Sce.Atf.Controls.PropertyEditing
         /// Constructs BoundedFloatEditor using the given arguments</summary>
         /// <param name="min">Min value</param>
         /// <param name="max">Max value</param>
-        public BoundedFloatEditor(int min = 0, int max = 0)
+        public BoundedFloatEditor(float min = 0, float max = 0)
         {
             if (min >= max)
                 throw new ArgumentOutOfRangeException("min must be less than max");
